requirements.txt
groq
langchain
streamlit
python-dotenv
PyPDF2

-e .